thanks to na_game_tool 0.4  more video formats are now supported

+++ src/Program.js
@@ -67,6 +67,7 @@ export class Program
 			// flags & execution
 			args             : {type : "function", length : [0, 1]},
 			allowDupOut      : {type : "boolean"},
+			allowLargeOutput : {type : "boolean"},
 			bin              : {types : ["function", "string"]},
 			bruteFlags       : {type : Object},
 			chain            : {types : ["function", "string"]},
@@ -267,7 +268,7 @@ export class Program
 				const newFileRel = path.relative(f.outDir.absolute, newFilePath);
 				const newFile = await DexFile.create({root : f.root, absolute : newFilePath});
 
-				if(newFile.size>xu.GB && f.input.size<(xu.GB*0.20))
+				if(!this.allowLargeOutput && newFile.size>xu.GB && f.input.size<(xu.GB*0.20))
 				{
 					xlog.warn`Program ${fg.orange(this.programid)} deleting file that is ${newFile.size.bytesToSize()} as it exceeds arbitrary 1GB size sanity check limit from source file that isn't very big: ${fg.green(newFileRel)}`;
 					await fileUtil.unlink(newFilePath);
